+++ node/middlewares/verifyOrderFormAffiliation.ts
@@ -1,20 +1,22 @@
 import type { App } from '@vtex/clients'
 
-import { APP_CUSTOM_DATA } from '../utils/constants'
+import { APP_CUSTOM_DATA, SUCCESS } from '../utils/constants'
 
 export async function verifyOrderFormAffiliation(
-  { state }: UserLoginEventContext,
+  ctx: UserLoginContext,
   next: () => Promise<unknown>
 ) {
   const {
     orderForm: { customData },
-  } = state
+  } = ctx.state
 
   const affiliateApp = customData?.customApps?.find(
     (app: App) => app.id === APP_CUSTOM_DATA.id
   )
 
   if (affiliateApp?.fields?.affiliateId) {
+    ctx.status = SUCCESS
+
     return
   }