+++ CHANGELOG.md
@@ -11,6 +11,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Service middleware unit tests after changes on implementation
 
+## [0.18.0] - 2021-12-22
+
+### Added
+
+- vtex.affiliates-order-form-spy@0.x as a dependency
+
+### Updated
+
+- verifyUserAffiliateLead to handle a service call instead of an event
+
 ## [0.17.0] - 2021-12-21
 
 ### Added